--- src/components/Popper/index.ts
@@ -1,1 +1,0 @@
-export * from './usePopper';
