+++ src/components/Select/Select.tsx
@@ -4,13 +4,13 @@ import * as React from 'react';
 import { HiChevronDown, HiX } from 'react-icons/hi';
 import styled from 'styled-components';
 import { sxMixin, SxProps, useComponentStyles } from '../../system';
+import { usePopper } from '../../hooks';
 import { LiteralUnion } from '../../types';
 import { mergeRefs } from '../../util';
 import { Box } from '../Box';
 import { Flex } from '../Flex';
 import { Input } from '../Input';
 import { MenuItem } from '../Menu';
-import { usePopper } from '../Popper';
 import { Spinner } from '../Spinner';
 import { Stack } from '../Stack';
 import { Text } from '../Text';
